# .github/api-contract/chat_app_server_rs.openapi.yaml
...
openapi: 3.0.3
info:
  title: Chatos Main Backend API
  version: 0.1.0-bootstrap
  description: 'Bootstrap OpenAPI contract for chat_app_server_rs.

    This is an initial skeleton and will be expanded incrementally.

    '
servers:
- url: http://localhost:3997
components:
  schemas:
    StandardResponse:
      type: object
      description: Generic API success envelope placeholder.
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          description: Domain-specific payload.
          nullable: true
      additionalProperties: true
paths:
...
  /api/task-manager/tasks/{task_id}:
    patch:
      summary: Update task
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
      operationId: patchApiTaskManagerTasksByTaskId
...
  /api/task-manager/tasks/{task_id}/complete:
    post:
      summary: Complete task
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
      operationId: postApiTaskManagerTasksByTaskIdComplete
...